Implementation of server identification challenge upon connect.
A client now receives a challenge from the server as soon as it
connects.  The challenge string looks like:
[size]:[nonce]:[serverstring]:[protocolversion]

In the current implementation, the nonce is not used for salting of a
md5 hash of the user password.  In fact, the challenge string is ignored
by the Mapi library.  JDBC uses the challenge to decide whether it can
(and will) send a request for a specific database.