[ 1169869 ] Mserver --dbinit="module(pathfinder);" crashes
Date: 2005-03-24 16:40
Sender: boncz
Logged In: YES
user_id=591107

appears to be an old one; I only was the unlucky guy that added
some extra work to the initialization of the pathfinder module
(it now loads pf_support).

Thus, the MIL stack got deeper, calls THRhighwater() and
crashes, because the executing thread is not registered.
Previously, with a more shallow stack, THRhighwater was not
called, and thus it survived.

The problem is that the --dbinit option uses monet_eval() very
early; from within scheduleClient, not through the MIL job
queue.  Only when monetInterpreter() is eneterd the thread registers
as a Monet interpreter.

fix is to register early, and make THRnew resistent against
already registered threads calling it.