Fixed bugs in pf_read: the arguments of fread are (buffer, size, count, filep).  If you want to read until end-of-file and you don't know how much that is, you must use size == 1.  If you use the size of the buffer, and the data fits into the buffer, fread will return 0, indicating that it didn't do anything (since it coudn't read any objects of the requested size).  As a side effect, the buffer may or may not be filled with the data.  Also, the return value can then be used as a measure of how much data was read.  Given the return value, we can now zero-terminate the buffer.
On Windows, it so happened that in some cases the buffer was filled
with the data of the file plus one extra character.  This character
was probably the result of the \r\n -> \n translation that happens on
text files.  With these fixes this does not happen.

Also, pf_read allocates a buffer for the data, and since it is only
used once, it is trivial to free the data after use, so do so.