- evacuated "pathfinder/haskell" to new CVS module "pf-haskell" - evacuated "pathfinder/loader"  to new CVS module "pf-tools"
use "cvs -d:ext:<user>@cvs.sf.net:/cvsroot/monetdb co -P (pf-haskell|pf-tools)"
to check-out the new modules.

subscribe to the respective mailing list to get the check-in messages:
 http://lists.sourceforge.net/lists/listinfo/monetdb-pf-haskell
 http://lists.sourceforge.net/lists/listinfo/monetdb-pf-tools


- removed all code that was related to pathfinder/loader/pf-shred:
  + configure checks for BerkeleyDB
  + BerkeleyDB and/or pf-shred related documentation

- changes are wrapped in CVS tags
  "evacuating_haskell_loader_before" & "evacuating_haskell_loader_after".


NOTE:
pf-tools/loader will not compile as it is now.
To get it compiling (independent of pathfinder?)
the Makefile(s) and maybe also the code need to be fixed
and/or "pf-tools" need to get its own configure setup
(derived from pathfinder).