some more changes to the dead-code eliminator: - put back in comments in the generated MIL (Jan R request) - removed pos parameter passing and result extraction from method calls (milprint_summer) - it now removes empty else blocks and empty sequential blocks - added simple code motion - added some more technical documentation - introduced mark_pos, that checks for a fake_project(iter) and if so does a mark(1) iso mark_grp
remark:
 this fake project stuff really is getting ugly; we should just use constants and an extent!

question (to Jan):
 why do we do so often:

	iter.tunique().mark().reverse

 to get all unique iters, while we have those ready in loop00x???



This code motion moves statements into the next conditional in both branches
(if- and *always* else-, because if the else- does not exist it is introduced)
if that makes the statements disappear in one of those.

code motion is still simple because:
- the analysis of whether a statement is moveable is very strict (all subsequent
  statements that overwrite one of its inputs or its outputs invalidate it).
  The more accurate check is to only check against those statements whose
  computation is required for calculating the boolean if-condition.
- statements are only moved into the directly following conditional, not
  further down the statement stream.

I could potentially enhance code motion in this direction, but for the
moment this simple strategy handles those cases that I know of.

These last statements cause a clean "grep pos" on the XMark queries,
except Q19 which uses order by (and thus needs pos).