added new test that compare the current semantics of insert() and append() to their new/desired semantics: *.old.* show the current/old behaviour, *.new.* show the desired/new behaviour.
for now, only insert and append of atoms are tests;
insert and append of BATs will follow soon.

Note/Background:

The tests also depict some of the bugs when
inserting in and/or appending to void|void BATs;
we will use the fixing of these bugs as opportunity
to clean-up the semantics of insert() and append()
(especially on VOID and NIL columns) as scetched in the
stable output of the *.new.* tests.