For RPC modules: add a string, 'rpc_uri', in PFpnode_t to hold the URI given by the 'at' hint, when necessary.  If a function is declared in an RPC module or if a function call is calling an RPC function, 'rpc_uri' contains the URI string; otherwise, it is a NULL pointer.
Later on, the RPC sender needs this URI to send to the RPC receiver so
that the receiver can know from where the module kan be loaded.
Currently, this URI is incorrectly given as the module namespace.
Henceforce, the module namespace and the URI of the module are given
separately, so that the RPC receiver can do a sanity check on the module
namespace it receives.

--TODO: implement the sanity check.